move Memory Enum to plan feature
