+++ conoha_client/features/plan/__init__.py
@@ -1,7 +1,12 @@
 """VM Plan feature."""
 
 from .cli import vm_plan_cli
-from .domain import VMPlan
+from .domain import Memory, VMPlan
 from .repo import first_vmplan_by
 
-__all__ = ["vm_plan_cli", "VMPlan", "first_vmplan_by"]
+__all__ = [
+    "vm_plan_cli",
+    "VMPlan",
+    "Memory",
+    "first_vmplan_by",
+]
