+++ conoha_client/features/plan/domain.py
@@ -1,6 +1,6 @@
 """VM Plan Domain."""
-from __future__ import annotations
 
+from enum import Enum
 from uuid import UUID
 
 from pydantic import BaseModel, Field
@@ -14,3 +14,34 @@ class VMPlan(BaseModel, frozen=True):
     mem_mb: int = Field(alias="ram", description="RAMのメモリ容量(MB)")
     n_core: int = Field(alias="vcpus", description="CPUのコア数")
     disk_gb: int = Field(alias="disk", description="SSDブートディスク容量(GB)")
+
+
+# 数字始まりの変数名にできなので単位をprefixにした
+class Memory(str, Enum):
+    """VMのメモリ容量."""
+
+    MG512 = "0.5"
+    GB1 = "1"
+    GB2 = "2"
+    GB4 = "4"
+    GB8 = "8"
+    GB16 = "16"
+    GB32 = "32"
+    GB64 = "64"
+
+    @property
+    def expression(self) -> str:
+        """Flavor(VMプラン)を特定するための表現."""
+        if self.is_smallest():
+            return "m512d"
+        return f"m{self.value}d"
+
+    def is_smallest(self) -> bool:
+        """最小のメモリか."""
+        return self == Memory.MG512
+
+    def is_match(self, img_name: str) -> bool:
+        """valueがイメージ名に含まれているか."""
+        if self.is_smallest():
+            return "30gb" in img_name
+        return "100gb" in img_name
