+++ conoha_client/features/plan/errors.py
@@ -1,0 +1,5 @@
+"""VM Plan errors."""
+
+
+class NotFoundFlavorError(Exception):
+    """Flavorが見つからないはずがない."""
